+++ src/main/java/com/supermartijn642/trashcans/screen/UltimateTrashCanContainer.java
@@ -6,7 +6,7 @@ import com.supermartijn642.trashcans.filter.ItemFilter;
 import com.supermartijn642.trashcans.filter.LiquidTrashCanFilters;
 import net.minecraft.core.BlockPos;
 import net.minecraft.world.entity.player.Player;
-import net.minecraft.world.inventory.ClickType;
+import net.minecraft.world.inventory.ContainerInput;
 import net.minecraft.world.item.ItemStack;
 
 /**
@@ -64,7 +64,7 @@ public class UltimateTrashCanContainer extends TrashCanContainer {
     }
 
     @Override
-    public void clicked(int slotId, int dragType, ClickType clickTypeIn, Player player){
+    public void clicked(int slotId, int dragType, ContainerInput input, Player player){
         if(!this.validateObjectOrClose())
             return;
 
@@ -89,7 +89,7 @@ public class UltimateTrashCanContainer extends TrashCanContainer {
             this.object.dataChanged();
             return;
         }
-        super.clicked(slotId, dragType, clickTypeIn, player);
+        super.clicked(slotId, dragType, input, player);
     }
 
     @Override