+++ changelog.md
@@ -1,2 +1,4 @@
+## Update to Minecraft 26.1
+
 ### Trash Cans 1.0.18
 - Initial release of Trash Cans for Minecraft 1.21.11